Smart-Water-Meter-System
Currently the municipal corporation are using analog meter which is not able to monitor online. The method adapted for water meters are to detection and control of water leakage, water supply monitoring and controlling system with theft identification. Meters count water consumption and at the end of it, send the data to the cloud server. The most significant advantage is that when people look at their own consumption data, they start taking conscious steps to optimize their consumption and bring in water and cost savings. The smart water meter will be digitized and automated, high accuracy maintained by decreasing human efforts. The whole distribution is in digitally can be continuously monitored by smart water meter system and cut off the water supply, in case a consumer exceeds the overdue limit.